feature: limit resolving v2 + tests (#15)
feature: migration to v2 of limits resolver + tests

+++ backend/src/main/kotlin/dev/holdbetter/innerApi/model/DayLimit.kt
@@ -8,5 +8,5 @@ data class DayLimit(
     val plannedDayLimit: Int,
     val firstMatchStartOrDefault: Instant,
     val remainedDayLimit: Int,
-    val updateRate: Int
+    val updateRate: Double
 )
